embedded_uniffi_bindgen

Module config_supplier

Structs

CargoLibrary ğŸ”’

CargoPackage ğŸ”’

CargoToml ğŸ”’

CargoTomlFile ğŸ”’

CargoWorkspace ğŸ”’

NoCargoConfigSupplier

Enums

CargoDependency ğŸ”’

Statics

CRATE_MAP ğŸ”’

Functions

find_other_cargo_toml_paths ğŸ”’

Process Cargo.toml data and return all crate paths referenced in it

find_workspace_crates ğŸ”’

Find all crates in this workspace and return a map of crate_name => crate_root_path

find_workspace_toml ğŸ”’

Find the workspace Cargo.toml file.

join ğŸ”’

join_and_glob ğŸ”’