embedded_uniffi_bindgen::uniffi_bindgen

Enum Commands

enum Commands {
    Generate {
        language: Vec<TargetLanguage>,
        out_dir: Option<Utf8PathBuf>,
        no_format: bool,
        config: Option<Utf8PathBuf>,
        lib_file: Option<Utf8PathBuf>,
        library_mode: bool,
        crate_name: Option<String>,
        source: Utf8PathBuf,
    },
    Scaffolding {
        out_dir: Option<Utf8PathBuf>,
        no_format: bool,
        udl_file: Utf8PathBuf,
    },
    Pipeline(PipelineArgs),
}

Variants

Generate

Generate foreign language bindings

Fields

language: Vec<TargetLanguage>

Foreign language(s) for which to build bindings.

out_dir: Option<Utf8PathBuf>

Directory in which to write generated files. Default is same folder as .udl file.

no_format: bool

Do not try to format the generated bindings.

config: Option<Utf8PathBuf>

Path to optional uniffi config file. This config is merged with the uniffi.toml config present in each crate, with its values taking precedence.

lib_file: Option<Utf8PathBuf>

Extract proc-macro metadata from a native lib (cdylib or staticlib) for this crate.

library_mode: bool

Pass in a cdylib path rather than a UDL file

crate_name: Option<String>

When --library is passed, only generate bindings for one crate. When --library is not passed, use this as the crate name instead of attempting to locate and parse Cargo.toml.

source: Utf8PathBuf

Path to the UDL file, or cdylib if library-mode is specified

Scaffolding

Generate Rust scaffolding code

Fields

out_dir: Option<Utf8PathBuf>

Directory in which to write generated files. Default is same folder as .udl file.

no_format: bool

Do not try to format the generated bindings.

udl_file: Utf8PathBuf

Path to the UDL file.

Pipeline(PipelineArgs)

Inspect the bindings render pipeline